mv some galaxy spec archive bits to content_archive
