+++ ansible_galaxy/archive.py
@@ -9,21 +9,16 @@ import os
 import yaml
 
 
+from ansible_galaxy import display
 from ansible_galaxy import exceptions
-from ansible_galaxy.models.content import CONTENT_TYPE_DIR_MAP, CONTENT_TYPES
+from ansible_galaxy.models import content
 from ansible_galaxy.models import content_repository
 
 log = logging.getLogger(__name__)
 
 # pass in list of tarinfo of paths to extract
 # pass in a map of tar member paths -> dest paths, built separately?
-#  (based on content_type and CONTENT_TYPE_DIR_MAP etc)
-
-
-def default_display_callback(*args, **kwargs):
-    # log.debug('args=%s, kwargs=%s', args, kwargs)
-
-    print(''.join(args))
+#  (based on content_type and content.CONTENT_TYPE_DIR_MAP etc)
 
 # for plugins and everything except roles
 # extract_content_by_content_type(content_type, base_path=None)
@@ -32,120 +27,6 @@ def default_display_callback(*args, **kwargs):
 # extract_content_by_role_name(role_name)
 
 # def content_type_match(content_type, member_path):
-
-
-# TODO:
-
-# TODO: better place to define?
-META_MAIN = os.path.join('meta', 'main.yml')
-GALAXY_FILE = 'ansible-galaxy.yml'
-APB_YAML = 'apb.yml'
-
-
-class ContentArchiveMetadataFiles(object):
-    def __init__(self):
-        self.meta_main_yml = None
-        self.meta_main_parent_dir = None
-
-        self.galaxy_file = None
-        self.apb_yaml = None
-
-        # FIXME: doesnt belong here
-        self.archive_parent_dir = None
-
-
-# TODO/FIXME: try to make sense of this and find_archive_parent_dir
-def find_archive_metadata(archive_members):
-    '''Try to find the paths to the archives meta data files
-
-    Aka, meta/main.yml or ansible-galaxy.yml.
-
-    Also, while we are at it, try to find the archive parent
-    dir.'''
-
-    meta_file = None
-    galaxy_file = None
-
-    meta_parent_dir = None
-    archive_parent_dir = None
-
-    apb_yaml_file = None
-
-    for member in archive_members:
-        if fnmatch.fnmatch(member.name, '*/%s' % APB_YAML):
-            log.debug('apb.yml member: %s', member)
-            apb_yaml_file = member.name
-
-        if META_MAIN in member.name or GALAXY_FILE in member.name:
-            # Look for parent of meta/main.yml
-            # Due to possibility of sub roles each containing meta/main.yml
-            # look for shortest length parent
-            meta_parent_dir = os.path.dirname(os.path.dirname(member.name))
-            if not meta_file:
-                archive_parent_dir = meta_parent_dir
-                if GALAXY_FILE in member.name:
-                    galaxy_file = member
-                else:
-                    meta_file = member
-            else:
-                # self.log.debug('meta_parent_dir: %s archive_parent_dir: %s len(m): %s len(a): %s member.name: %s',
-                #               meta_parent_dir, archive_parent_dir,
-                #               len(meta_parent_dir),
-                #               len(archive_parent_dir),
-                #               member.name)
-                if len(meta_parent_dir) < len(archive_parent_dir):
-                    archive_parent_dir = meta_parent_dir
-                    meta_file = member
-                    if GALAXY_FILE in member.name:
-                        galaxy_file = member
-                    else:
-                        meta_file = member
-
-    log.debug('apb_yaml_file: %s', apb_yaml_file)
-    # FIXME: return a real type/object for archive metadata
-    return (meta_file,
-            meta_parent_dir,
-            galaxy_file,
-            apb_yaml_file)
-
-
-def find_archive_parent_dir(archive_members, content_type, content_dir):
-    # archive_parent_dir wasn't found when checking for metadata files
-    archive_parent_dir = None
-
-    shortest_dir = None
-    for member in archive_members:
-        # This is either a new-type Galaxy Content that doesn't have an
-        # ansible-galaxy.yml file and the type desired is specified and
-        # we check parent dir based on the correct subdir existing or
-        # we need to just scan the subdirs heuristically and figure out
-        # what to do
-        member_dir = os.path.dirname(os.path.dirname(member.name))
-        shortest_dir = shortest_dir or member_dir
-
-        if len(member_dir) < len(shortest_dir):
-            shortest_dir = member_dir
-
-        if content_type != "all":
-            if content_dir and content_dir in member.name:
-                archive_parent_dir = os.path.dirname(member.name)
-                return archive_parent_dir
-        else:
-            for plugin_dir in CONTENT_TYPE_DIR_MAP.values():
-                if plugin_dir in member.name:
-                    archive_parent_dir = os.path.dirname(member.name)
-                    return archive_parent_dir
-
-    if content_type not in CONTENT_TYPES:
-        log.debug('did not find a content_dir or plugin_dir, so using shortest_dir %s for archive_parent_dir', shortest_dir)
-        return shortest_dir
-
-    # TODO: archive format exception?
-    msg = "No content metadata provided, nor content directories found for content_type: %s" % \
-        content_type
-    log.debug('content_type: %s', content_type)
-    log.debug('content_dir: %s', content_dir)
-    raise exceptions.GalaxyClientError(msg)
 
 
 # FIXME: mv to AnsibleGalaxyMetadata
@@ -201,7 +82,7 @@ def find_content_type_subdirs(tar_file_members):
     parent dir, which will be self.content_meta.name (we don't want it as it's
     not needed for plugin types. First make sure the length of
     that split and drop of parent dir is length > 1 and verify
-    that the subdir is infact in CONTENT_TYPE_DIR_MAP.values()
+    that the subdir is infact in content.CONTENT_TYPE_DIR_MAP.values()
 
     This should give us a list of valid content type subdirs
     found heuristically within this Galaxy Content repo'''
@@ -209,7 +90,7 @@ def find_content_type_subdirs(tar_file_members):
         os.path.join(m.name.split(os.sep)[1:])[0]
         for m in tar_file_members
         if len(os.path.join(m.name.split(os.sep)[1:])) > 1
-        and os.path.join(m.name.split(os.sep)[1:])[0] in CONTENT_TYPE_DIR_MAP.values()
+        and os.path.join(m.name.split(os.sep)[1:])[0] in content.CONTENT_TYPE_DIR_MAP.values()
     ]
 
     # uniq and order
@@ -261,7 +142,7 @@ def filter_members_by_content_type(tar_file_obj,
                       if tar_info_content_name_match(tar_file_member,
                                                      "",
                                                      # self.content_meta.name,
-                                                     content_path=CONTENT_TYPE_DIR_MAP.get(content_type))]
+                                                     content_path=content.CONTENT_TYPE_DIR_MAP.get(content_type))]
 
     # everything for roles
     if content_archive_type == 'role':
@@ -371,7 +252,7 @@ def extract_by_content_type(tar_file_obj,
     elog.debug('extract_to_path=%s', extract_to_path)
     elog.debug('content_meta=%s', content_meta)
 
-    display_callback = display_callback or default_display_callback
+    display_callback = display_callback or display.display_callback
     files_to_extract = files_to_extract or []
 
     # if content_type != "role" and content_type not in self.NO_META:
@@ -386,7 +267,7 @@ def extract_by_content_type(tar_file_obj,
 
     # append the content_dir if we have one
     content_path = os.path.join(extract_to_path,
-                                CONTENT_TYPE_DIR_MAP.get('install_content_type', content_meta.content_dir or ''))
+                                content.CONTENT_TYPE_DIR_MAP.get('install_content_type', content_meta.content_dir or ''))
     if content_meta.content_sub_dir:
         log.debug('content_sub_path=%s', content_meta.content_sub_dir)
         content_path = os.path.join(content_path, content_meta.content_sub_dir or '')
@@ -426,8 +307,8 @@ def extract_by_content_type(tar_file_obj,
                     plugin_found = parent_dir.lstrip(content_meta.name)
 
             # secondary dir (roles/, callback_plugins/) is a match for the content_type
-            elif len(parts_list) > 1 and parts_list[1] == CONTENT_TYPE_DIR_MAP.get(content_meta.content_type):
-                plugin_found = CONTENT_TYPE_DIR_MAP.get(content_meta.content_type)
+            elif len(parts_list) > 1 and parts_list[1] == content.CONTENT_TYPE_DIR_MAP.get(content_meta.content_type):
+                plugin_found = content.CONTENT_TYPE_DIR_MAP.get(content_meta.content_type)
 
             # log.debug('plugin_found1: %s', plugin_found)
             # if not plugin_found:
