+++ ansible_galaxy/content_archive.py
@@ -1,13 +1,18 @@
+import fnmatch
 import logging
 import os
 import tarfile
 
-from ansible_galaxy import archive
 from ansible_galaxy import exceptions
 from ansible_galaxy.models import content
 from ansible_galaxy.models import content_archive
 
 log = logging.getLogger(__name__)
+
+# TODO: better place to define?
+META_MAIN = os.path.join('meta', 'main.yml')
+GALAXY_FILE = 'ansible-galaxy.yml'
+APB_YAML = 'apb.yml'
 
 
 def detect_content_archive_type(archive_path, archive_members):
@@ -41,6 +46,100 @@ def detect_content_archive_type(archive_path, archive_members):
     return None
 
 
+# TODO/FIXME: try to make sense of this and find_archive_parent_dir
+def find_archive_metadata(archive_members):
+    '''Try to find the paths to the archives meta data files
+
+    Aka, meta/main.yml or ansible-galaxy.yml.
+
+    Also, while we are at it, try to find the archive parent
+    dir.'''
+
+    meta_file = None
+    galaxy_file = None
+
+    meta_parent_dir = None
+    archive_parent_dir = None
+
+    apb_yaml_file = None
+
+    for member in archive_members:
+        if fnmatch.fnmatch(member.name, '*/%s' % APB_YAML):
+            log.debug('apb.yml member: %s', member)
+            apb_yaml_file = member.name
+
+        if META_MAIN in member.name or GALAXY_FILE in member.name:
+            # Look for parent of meta/main.yml
+            # Due to possibility of sub roles each containing meta/main.yml
+            # look for shortest length parent
+            meta_parent_dir = os.path.dirname(os.path.dirname(member.name))
+            if not meta_file:
+                archive_parent_dir = meta_parent_dir
+                if GALAXY_FILE in member.name:
+                    galaxy_file = member
+                else:
+                    meta_file = member
+            else:
+                # self.log.debug('meta_parent_dir: %s archive_parent_dir: %s len(m): %s len(a): %s member.name: %s',
+                #               meta_parent_dir, archive_parent_dir,
+                #               len(meta_parent_dir),
+                #               len(archive_parent_dir),
+                #               member.name)
+                if len(meta_parent_dir) < len(archive_parent_dir):
+                    archive_parent_dir = meta_parent_dir
+                    meta_file = member
+                    if GALAXY_FILE in member.name:
+                        galaxy_file = member
+                    else:
+                        meta_file = member
+
+    log.debug('apb_yaml_file: %s', apb_yaml_file)
+    # FIXME: return a real type/object for archive metadata
+    return (meta_file,
+            meta_parent_dir,
+            galaxy_file,
+            apb_yaml_file)
+
+
+def find_archive_parent_dir(archive_members, content_type, content_dir):
+    # archive_parent_dir wasn't found when checking for metadata files
+    archive_parent_dir = None
+
+    shortest_dir = None
+    for member in archive_members:
+        # This is either a new-type Galaxy Content that doesn't have an
+        # ansible-galaxy.yml file and the type desired is specified and
+        # we check parent dir based on the correct subdir existing or
+        # we need to just scan the subdirs heuristically and figure out
+        # what to do
+        member_dir = os.path.dirname(os.path.dirname(member.name))
+        shortest_dir = shortest_dir or member_dir
+
+        if len(member_dir) < len(shortest_dir):
+            shortest_dir = member_dir
+
+        if content_type != "all":
+            if content_dir and content_dir in member.name:
+                archive_parent_dir = os.path.dirname(member.name)
+                return archive_parent_dir
+        else:
+            for plugin_dir in content.CONTENT_TYPE_DIR_MAP.values():
+                if plugin_dir in member.name:
+                    archive_parent_dir = os.path.dirname(member.name)
+                    return archive_parent_dir
+
+    if content_type not in content.CONTENT_TYPES:
+        log.debug('did not find a content_dir or plugin_dir, so using shortest_dir %s for archive_parent_dir', shortest_dir)
+        return shortest_dir
+
+    # TODO: archive format exception?
+    msg = "No content metadata provided, nor content directories found for content_type: %s" % \
+        content_type
+    log.debug('content_type: %s', content_type)
+    log.debug('content_dir: %s', content_dir)
+    raise exceptions.GalaxyClientError(msg)
+
+
 def load_archive(archive_path):
     archive_parent_dir = None
 
@@ -58,7 +157,7 @@ def load_archive(archive_path):
 
     # next find the metadata file
     (meta_file, meta_parent_dir, galaxy_file, apb_yaml_file) = \
-        archive.find_archive_metadata(members)
+        find_archive_metadata(members)
 
     archive_type = detect_content_archive_type(archive_path, members)
     log.debug('archive_type: %s', archive_type)
